refactor: modularização das classes
